// src/app/tabs/shared/models/estidama-chart.model.ts
export class EstidamaChart {
    MeterTypeName?: string;
    MeterId?: string;
    ClientId?: string;
    ReportType?: string;    
    GroupId?: string;
    FromYear?: string;
    ToYear?: string;
    FromDate?: string;
    ToDate?: string;
    FromMonth?: string;
    ToMonth?: string;
    BlnCompare?: boolean;
}